// apps/web/src/lib/api/generated/types/groupMovieUpdateDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Movies App API
 * REST API for Movies application

**Links:**
- [OpenAPI JSON](/api/docs/json)
- [OpenAPI YAML](/api/docs/yaml)
 * OpenAPI spec version: 1.0.0
 */
import type { GroupMovieUpdateDtoStatus } from './groupMovieUpdateDtoStatus';

export interface GroupMovieUpdateDto {
	/** Movie status in group */
	status?: GroupMovieUpdateDtoStatus;
	/** Planned watch date */
	plannedDate?: string;
	/** Watched date */
	watchedDate?: string;
}